refactor: SettingsModal の 17 セクションを 1 セクション 1 コンポーネントに分割 (#1126)
907 行の SettingsModal.vue を、src/components/settings/ の 17 ファイルに分ける。
セクションは互いに状態を共有していないので、それぞれ自分の composable か自分の
prop/emit 1 組だけを持つ。殻に残るのは dialog のフレーム・キーボード操作・
セクションの並び順・props/emits の転送で 147 行。

props/emits の口は変えていない。SettingsModal.spec.ts (369 行) は mount から
ツリー全体を引いているので、emit を親まで転送すれば無改変で通る。

あわせて重複 3 件:
- SettingsListRow.vue — 削除ボタン付きの行 (4 箇所で同一だった li 枠 + close ボタン)
- SettingsStepper.vue — −/値/+ (font size と scroll speed でほぼ同一だった 25 行)
- useSavedListMirror — 保存値のミラー + 丸ごと差し替え emit (4 箇所)

見出しクラスの SECTION_HEADING は settings/sectionClasses.ts へ移動。<ul> 側の
6 連 utility も SETTINGS_LIST として同じ場所に。

描画が変わっていないことは、分割前のコンポーネントを一時的に併置して同じ props で
両方 mount し html() を diff して確認した (9 通りの条件、構造差ゼロ)。残った差は
whitespace だけ — 詳細は plans/refactor-1126-settings-sections.md。
